# openapi.yaml
openapi: 3.0.3
info:
  title: OZU AI Project Team API
  version: 0.1.0
servers:
  - url: https://api.ozuai.tech
paths:
...
  /idea/{id}:
    get:
      summary: Idea detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Idea detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int32
                  title:
                    type: string
                  description:
                    type: string
                  status:
                    $ref: '#/components/schemas/IdeaStatus'
                  owners:
                    type: array
                    description: Idea owners
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: int32
                        email:
                          type: string
                          format: email
                        name:
                          type: string
                          nullable: true
                        surname:
                          type: string
                          nullable: true
                        teamMember:
                          type: boolean
                      required: [id, email]
                  teamMembers:
                    type: array
                    description: Team members working on the idea
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: int32
                        email:
                          type: string
                          format: email
                        name:
                          type: string
                          nullable: true
                        surname:
                          type: string
                          nullable: true
                        teamMember:
                          type: boolean
                      required: [id, email]
                  presentations:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: int32
                        date:
                          type: string
                          format: date-time
                        done:
                          type: boolean
                      required: [id, date, done]
                  applicationId:
                    type: integer
                    format: int32
                    nullable: true
                required:
                  - id
                  - title
                  - description
                  - status
                  - owners
        '404':
          description: Idea not found

  # ===================== APPLICATIONS =====================
...
  /presentation:
    get:
      summary: Presentation list
      parameters:
        - in: query
          name: current
          schema:
            type: boolean
          required: false
          description: If true, returns only upcoming/current presentations (not yet done).
      responses:
        '200':
          description: Presentation list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int32
                    title:
                      type: string
                    date:
                      type: string
                      format: date-time
                    done:
                      type: boolean
                    own:
                      type: boolean
                      description: Is the requesting user in the presenting team?
                  required: [id, title, date, done, own]
  /presentation/{id}:
    get:
      summary: Presentation detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Presentation detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  idea:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int32
                      title:
                        type: string
                    required: [id, title]
                  title:
                    type: string
                  summary:
                    type: string
                  date:
                    type: string
                    format: date-time
                  location:
                    type: string
                  done:
                    type: boolean
                  owners:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          format: int32
                        email:
                          type: string
                          format: email
                        name:
                          type: string
                          nullable: true
                        surname:
                          type: string
                          nullable: true
                        teamMember:
                          type: boolean
                      required: [id, email]
                  feedbacks:
                    type: array
                    items:
                      type: object
                      properties:
                        content:
                          type: string
                        user:
                          type: object
                          properties:
                            id:
                              type: integer
                              format: int32
                            name:
                              type: string
                              nullable: true
                            surname:
                              type: string
                              nullable: true
                            teamMember:
                              type: boolean
                          required: [id]
                        date:
                          type: string
                          format: date-time
                required:
                  - idea
                  - title
                  - summary
                  - date
                  - location
                  - done
        '404':
          description: Presentation not found

  # ===================== USERS =====================
  /user/{id}:
    get:
      summary: User profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: User basic information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int32
                  email:
                    type: string
                    format: email
                  name:
                    type: string
                    nullable: true
                  surname:
                    type: string
                    nullable: true
                  studentNumber:
                    type: string
                    nullable: true
                  teamMember:
                    type: boolean
                    description: Is there a TeamMember record?
                required:
                  - id
                  - email
        '404':
          description: User not found
  /user/{id}/team-member:
    get:
      summary: User's team member profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Team member profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int32
                  workAreas:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkArea'
                  photoURL:
                    type: string
                    nullable: true
                  bio:
                    type: string
                  github:
                    type: string
                    nullable: true
                  linkedin:
                    type: string
                    nullable: true
                  extraLinks:
                    type: array
                    items:
                      type: string
                  applicationId:
                    type: integer
                    format: int32
                    nullable: true
                required:
                  - id
                  - workAreas
                  - bio
        '404':
          description: Team member profile not found

  # ===================== AUTH =====================
...
  /auth/signin:
    post:
      summary: Sign in with code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
              required: [email, code]
      responses:
        '200':
          description: Sign in successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  role:
                    $ref: '#/components/schemas/UserRole'
                  accessToken:
                    type: string
                required: [role, accessToken]
...
components:
  schemas:
...
    UserRole:
      type: string
      enum:
        - STUDENT
        - MEMBER
        - ADMIN
...
    WorkArea:
      type: string
      enum:
        - FRONTEND
        - BACKEND
        - DESIGN
        - MOBILE
        - MANAGEMENT
...
    IdeaStatus:
      type: string
      enum:
        - DRAFT
        - PRESENTED
        - ACCEPTED
        - REJECTED
...